requirements.txt
aiohttp
aiolimiter
pandas
google-cloud-bigquery
python-dotenv
pytest
pytest-mock
pytest-asyncio
playwright

gql[aiohttp]
requests

tweepy
praw
pyyaml
discord.py
telethon
transformers
nltk
feedparser
trafilatura
bertopic
sentence-transformers
hdbscan
umap-learn
openai
slack_sdk
scikit-learn

joblib
xgboost
tqdm
